Help with analyzer ignores (#50)
* Refactor initialize into smaller methods

Take the // Do Stuff commented blocks and turn them into _doStuff()
methods.

Makes the data flow between the various steps explicit and will make it
easier to add new steps.

* Add init step to check for analysis_options

Fixes #30

Search for an analysis_options in the pub package directory, or any
parent directory. If one exists, check for exclude 'bazel-*' and print a
help message if it is missing. If no analysis_options exists create a
default.

* Import packages_file with a prefix