Add a StepTimer class (#53)
- Add StepTimer which can time individual steps and print all the timing
  data
- Use StepTimer from the Initialize run method to reduce duplication

The StepTimer run() method is async, which means that some time could
get lost if other futures execute between the call sight and the next
event loop. This should be fine since the data we care about is how long
each step takes, and we don't care if the steps necessarily add up
exactly to the total time.