Drop stale unsupported concept replacesTransfomer (#156)
We don't currently support automatically reading pubspec.yaml
transformer config and dropping in a Builder in it's place. Adding this
support is likely more effort than the value we'd get out of it - so for
now drop the concept entirely.

Documentation for this is cleaned up in a separate commit.