Explicit error when srcs are duplicated in deps (#160)
Most builders expect to be able to do deeper resolution on their inputs
than what is possible with summaries. When a URI is resolvable through
both the file and a summary the summary ends up getting preferred.
Detect when this situation happens and throw a more explicit message
rather than defer until some resolution is attempted that doesn't work.
This will make the resolver behave more consistently and not silently
fallback to a less powerful mode in some cases.

It's possible that this will cause failures in places that otherwise
would have worked if the Builder wouldn't have needed the deeper
resolution.

One place where this comes up is `source_gen` where we require that
inputs with parts use non-summary resolution.