Use inputs before summaries in analysis context (#169)
In some cases a source file can show up in multiple targets. When a
source is also used in a dependency then resolving the file will use the
version from a summary, which is missing some information that is needed
for codegen. Reordering the URI resolvers give precedence to the
resolver that reads from the input.

This will make performance more sensitive to bugs in the skylark rules
that may cause more Dart files to show up as inputs rather than only in
summaries.