Use accumlated inputs/outputs
Previously any configuration with multiple builders in series needed to
behave as a perfect chain - the outputs from one phase needed to be
exactly the primary inputs to the next phase.

In build 0.9.1 the inputs are filtered to only those which will produce
an output, so we can continue to collect inputs and only the assets
which should be primary inputs will end up being used.