Make rootPackage private in BazelAssetReader
We will need to make an implementation of this class in a test and the
field does not match the field named `rootPackage` in the existing
`InMemoryAssetReader` from `build_test`. This field isn't used outside
of the class so make it private. Rename for consistency in case we ever
decide to make it public again.